Fixed the mistakes

+++ skuratov.danil/S1/sequence.hpp
@@ -14,12 +14,12 @@ namespace skuratov
       name_(name)
     {}
     const std::string& name() const { return name_; }
-    std::vector<int>& numbers() { return numbers_; }
-    const std::vector<int>& numbers() const { return numbers_; }
+    std::vector<size_t>& numbers() { return numbers_; }
+    const std::vector<size_t>& numbers() const { return numbers_; }
 
   private:
     std::string name_;
-    std::vector<int> numbers_;
+    std::vector<size_t> numbers_;
   };
 }
